1. remove unused file(origin skin file) 2. move singer from beatmap into skin
