+++ osu.Game.Rulesets.Karaoke/Skinning/Components/Singer.cs
@@ -1,0 +1,18 @@
+﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using osuTK.Graphics;
+
+namespace osu.Game.Rulesets.Karaoke.Skinning.Components
+{
+    public class Singer
+    {
+        public string Name { get; set; }
+
+        public string Romaji { get; set; }
+
+        public string EnglishName { get; set; }
+
+        public Color4 Color { get; set; }
+    }
+}
